Added Spawner Regulators
A Spawner can have regulators added to it.
They have two colliders, which one should be set as the *Inner* one, meaning it's closer to the spawner and should be triggered after the *Outer* one is triggered, when the player moves to the spawner (e.g. goes to a different area)